backend/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
google-generativeai
sqlalchemy
pydantic
